chore: removed the function and moved to helpers renamed v1.2
